unify announcements styles
