+++ src/library/Announcements/Header.tsx
@@ -3,14 +3,14 @@
 
 import { ButtonHelp, ButtonTertiary } from '@polkadot-cloud/react';
 import { useHelp } from 'contexts/Help';
-import { Wrapper } from './Wrapper';
-import type { StatsHeadProps } from './types';
+import { HeaderWrapper } from './Wrappers';
+import type { HeaderProps } from './types';
 
-export const StatsHead = ({ items }: StatsHeadProps) => {
+export const Header = ({ items }: HeaderProps) => {
   const { openHelp } = useHelp();
 
   return (
-    <Wrapper>
+    <HeaderWrapper>
       {items.map(({ label, value, button, helpKey }, i) => (
         <div key={`head_stat_${i}`}>
           <div className="inner">
@@ -33,6 +33,6 @@ export const StatsHead = ({ items }: StatsHeadProps) => {
           </div>
         </div>
       ))}
-    </Wrapper>
+    </HeaderWrapper>
   );
 };
