+++ src/library/Announcements/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export interface StatsHeadProps {
+export interface HeaderProps {
   items: PoolStatLabel[];
 }
