+++ src/library/GenerateNominations/index.tsx
@@ -18,7 +18,7 @@ import { SelectableWrapper } from 'library/List';
 import { SelectItems } from 'library/SelectItems';
 import { SelectItem } from 'library/SelectItems/Item';
 import { ValidatorList } from 'library/ValidatorList';
-import { Wrapper } from 'pages/Overview/NetworkSats/Wrappers';
+import { Wrapper } from './Wrapper';
 import { useStaking } from 'contexts/Staking';
 import { useFavoriteValidators } from 'contexts/Validators/FavoriteValidators';
 import { useActiveAccounts } from 'contexts/ActiveAccounts';
