+++ src/pages/Pools/Home/PoolStats/Announcements.tsx
@@ -11,7 +11,7 @@ import { useApi } from 'contexts/Api';
 import { useActivePool } from 'contexts/Pools/ActivePool';
 import { Announcement as AnnouncementLoader } from 'library/Loader/Announcement';
 import { useNetwork } from 'contexts/Network';
-import { Item } from './Wrappers';
+import { Item } from 'library/Announcements/Wrappers';
 
 export const Announcements = () => {
   const { t } = useTranslation('pages');
